- Profile comp. updated to new arch; move common files to /base folder; render-nip05,npub,stats logic extracted to functions
