+++ src/base/render-options.ts
@@ -1,0 +1,8 @@
+import { Theme } from "../common/types";
+
+export interface RenderOptions {
+  theme: Theme;
+  isLoading: boolean;
+  isError: boolean;
+  errorMessage: string;
+}
