+++ src/common/theme.ts
@@ -425,479 +425,3 @@ export const getProfileBadgeStylesLegacy = (theme: Theme) => {
   return `<style>${styles}</style>`;
 };
 
-// Function for NostrProfile styles
-export function getProfileStyles(theme: Theme) {
-  let variables = ``;
-
-  if (theme === 'dark') {
-    variables = `
-    --nstrc-profile-background: var(--nstrc-profile-background-dark, #000000);
-    --nstrc-profile-skeleton-min-hsl: var(--nstrc-profile-skeleton-min-hsl-dark, 200, 20%, 20%);
-    --nstrc-profile-skeleton-max-hsl: var(--nstrc-profile-skeleton-max-hsl-dark, 200, 20%, 30%);
-    --nstrc-profile-text-primary: var(--nstrc-profile-text-primary-dark, #ffffff);
-    --nstrc-profile-text-grey: var(--nstrc-profile-text-grey-dark, #666666);
-    --nstrc-profile-banner-placeholder-color: var(--nstrc-profile-banner-placeholder-color-dark, #222222);
-    --nstrc-profile-copy-foreground-color: var(--nstrc-profile-copy-foreground-color-dark, #CCC);
-    `;
-  } else {
-    variables = `
-    --nstrc-profile-background: var(--nstrc-profile-background-light, #f5f5f5);
-    --nstrc-profile-skeleton-min-hsl: var(--nstrc-profile-skeleton-min-hsl-light, 200, 20%, 80%);
-    --nstrc-profile-skeleton-max-hsl: var(--nstrc-profile-skeleton-max-hsl-light, 200, 20%, 95%);
-    --nstrc-profile-text-primary: var(--nstrc-profile-text-primary-light, #111111);
-    --nstrc-profile-text-grey: var(--nstrc-profile-text-grey-light, #808080);
-    --nstrc-profile-banner-placeholder-color: var(--nstrc-profile-banner-placeholder-color-light, #e5e5e5);
-    --nstrc-profile-copy-foreground-color: var(--nstrc-profile-copy-foreground-color-light, #222);
-    `;
-  }
-
-  return `
-  <style>
-  :host {
-      ${variables}
-
-      --nstrc-profile-accent: var(--nstrc-profile-accent, #ca077c);
-
-      /* Override follow button styles for profile context */
-      --nstrc-follow-btn-padding: 5px 8px !important;
-      --nstrc-follow-btn-font-size: 14px !important;
-      --nstrc-follow-btn-border-radius: 12px !important;
-      --nstrc-follow-btn-border-dark: 1px solid #DDDDDD !important;
-      --nstrc-follow-btn-border-light: 1px solid #DDDDDD !important;
-      --nstrc-follow-btn-horizontal-alignment: end !important;
-      --nstrc-follow-btn-min-height: auto !important;
-    }
-
-      .nostr-profile .skeleton {
-          animation: profile-skeleton-loading 0.5s linear infinite alternate;
-      }
-
-      @keyframes profile-skeleton-loading {
-          0% {
-              background-color: hsl(var(--nstrc-profile-skeleton-min-hsl));
-          }
-          100% {
-              background-color: hsl(var(--nstrc-profile-skeleton-max-hsl));
-          }
-      }
-
-    .nostr-profile {
-      -webkit-tap-highlight-color: transparent;
-      text-size-adjust: 100%;
-      font-weight: 400;
-      font-size: 18px;
-      line-height: 1.5;
-      text-rendering: optimizeLegibility;
-      overflow-wrap: break-word;
-      font-family: Nacelle, sans-serif;
-      -webkit-font-smoothing: antialiased;
-      box-sizing: border-box;
-      background-repeat: no-repeat;
-      min-height: 500px;
-      border: 1px solid #CCC;
-      border-radius: 5px;
-      background-color: var(--nstrc-profile-background);
-    }
-
-    #profile {
-      position: relative;
-      background-color: var(--nstrc-profile-background);
-      padding-bottom: 4px;
-    }
-
-    #profile_banner {
-      width: 100%;
-      height: 214px;
-      overflow: hidden;
-      cursor: pointer;
-    }
-
-    #profile_banner a {
-      outline: none;
-      color: var(--nstrc-profile-accent);
-    }
-
-    #profile_banner a img {
-      color: var(--nstrc-profile-accent);
-      max-width: 100%;
-      border-style: none;
-      display: block;
-      z-index: 22;
-      width: 100%;
-      height: 214px;
-      object-fit: cover;
-    }
-
-    #profile_banner .banner-placeholder {
-      width: 100%;
-      height: 214px;
-      background-color: var(--nstrc-profile-banner-placeholder-color);
-    }
-
-    .dp_container {
-      position: absolute;
-      top: 140px;
-      left: 15px;
-      overflow: hidden;
-    }
-
-    .avatar_container {
-      border: solid 4px var(--nstrc-profile-background);
-      border-radius: 50%;
-      background-color: var(--nstrc-profile-background);
-    }
-
-    .avatar_wrapper {
-      display: block;
-      min-height: 142px;
-    }
-
-    .xxl_avatar {
-      position: relative;
-      background-color: var(--nstrc-profile-background);
-      border-radius: 50%;
-      width: 142px;
-      height: 142px;
-    }
-
-    .backfill {
-      background-color: var(--nstrc-profile-background);
-      border-radius: 50%;
-      width: 142px;
-      height: 142px;
-    }
-
-    .backfill a {
-      outline: none;
-    }
-
-    .backfill a img {
-      max-width: 100%;
-      border-style: none;
-      display: block;
-      z-index: 22;
-      border-radius: 50%;
-      width: 100%;
-      height: 100%;
-      object-fit: cover;
-    }
-
-    .profile_actions {
-      height: 76px;
-      display: flex;
-      justify-content: end;
-      align-items: center;
-      padding: 0 18px;
-    }
-
-    .profile_data {
-      display: block;
-      margin-inline: 20px;
-      min-height: 52px;
-      margin-top: 14px;
-    }
-
-    .basic_info {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-    }
-
-    .name {
-      display: flex;
-      align-items: center;
-      justify-content: flex-start;
-      max-width: 60%;
-      height: 100%;
-    }
-
-    .name-text {
-      color: var(--nstrc-profile-text-primary);
-      font-size: 20px;
-      line-height: 1;
-      font-weight: 700;
-      text-overflow: ellipsis;
-      display: flex;
-      align-items: center;
-      vertical-align: baseline;
-    }
-
-    .verification-check {
-      width: 20px;
-      height: 20px;
-    }
-
-    .verification-icon {
-      width: 20px;
-      height: 20px;
-      position: relative;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      margin-inline: 4px;
-    }
-
-    ._whiteCheckL_lfbpc_49 {
-      width: 12px;
-      height: 12px;
-      top: 4px;
-      left: 4px;
-      border-radius: 50%;
-      background-color: #fff;
-      position: absolute;
-    }
-
-    ._verifiedIconPrimal_lfbpc_30 {
-      width: 100%;
-      height: 100%;
-      background: var(--nstrc-profile-accent);
-      mask: url(https://primal.net/assets/verified-84f47cd3.svg) no-repeat 0/100%;
-    }
-
-    .joined {
-      font-weight: 400;
-      font-size: 14px;
-      line-height: 16px;
-      text-align: right;
-      color: var(--nstrc-profile-text-grey);
-    }
-
-    .profile-ids {
-      display: flex;
-      flex-direction: column;
-      gap: 4px;
-      margin-top: 4px;
-      color: var(--nstrc-profile-text-grey);
-      font-size: 14px;
-      width: 100%;
-    }
-
-    .nip05-wrapper {
-      display: flex;
-      flex-direction: column;
-      align-items: flex-start;
-      color: var(--nstrc-profile-text-grey);
-      font-weight: 400;
-      font-size: 14px;
-      line-height: 16px;
-      margin-top: 2px;
-      margin-bottom: 16px;
-    }
-
-    .nip05-container {
-      color: var(--nstrc-profile-text-grey);
-      font-weight: 400;
-      font-size: 14px;
-      line-height: 16px;
-      display: flex;
-      align-items: center;
-    }
-
-    .nip05 {
-      color: var(--nstrc-profile-text-grey);
-      font-weight: 400;
-      font-size: 14px;
-      line-height: 16px;
-      width: 100%;
-      overflow: hidden;
-      display: flex;
-      gap: 5px;
-    }
-
-    .about {
-      margin-inline: 20px;
-      font-weight: 400;
-      font-size: 14px;
-      line-height: 20px;
-      color: var(--nstrc-profile-text-primary);
-    }
-
-    .links {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      margin-inline: 20px;
-      margin-block: 12px;
-    }
-
-    .website {
-      font-weight: 400;
-      font-size: 14px;
-      line-height: 20px;
-      display: flex;
-      align-items: center;
-    }
-
-    .website a {
-      font-weight: 400;
-      font-size: 14px;
-      line-height: 20px;
-      outline: none;
-      color: var(--nstrc-profile-accent);
-      max-width: 350px;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      word-wrap: normal;
-    }
-
-    .stats {
-      position: relative;
-      display: flex;
-      flex-direction: row;
-      flex-wrap: wrap;
-      justify-content: space-between;
-      align-items: center;
-      padding-inline: 8px;
-      border-radius: 0;
-      padding-top: 22px;
-      border-top: none;
-      background-color: var(--nstrc-profile-background);
-    }
-
-    .stat {
-      position: relative;
-      display: inline-block;
-      padding-inline: 14px;
-      padding-block: 2px;
-      border: none;
-      background: none;
-      width: fit-content;
-      height: 40px;
-      margin: 0 0 12px;
-    }
-
-    .stat-inner {
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      align-items: center;
-    }
-
-    .stat-inner .stat-value {
-      font-weight: 400;
-      font-size: 24px;
-      line-height: 24px;
-      color: var(--nstrc-profile-text-primary);
-    }
-
-    .stat-inner .stat-name {
-      font-weight: 400;
-      font-size: 14px;
-      line-height: 16px;
-      color: var(--nstrc-profile-text-grey);
-      text-transform: lowercase;
-    }
-
-    .error-container {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      gap: 20px;
-      min-height: 500px;
-    }
-
-    .error {
-      width: 35px;
-      height: 35px;
-      border-radius: 50%;
-      background-color: red;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      font-size: 20px;
-      color: #FFF;
-    }
-
-    .error-text {
-      color: red;
-      font-weight: bold;
-    }
-
-    .npub-container {
-      display: flex;
-      align-items: center;
-      gap: 4px;
-      margin-top: 2px;
-      width: 100%;
-    }
-
-    .npub {
-      color: #a2a2a2;
-      display: flex;
-      align-items: center;
-      gap: 4px;
-      word-break: break-all;
-      font-family: monospace;
-      font-size: 13px;
-    }
-    
-    .profile-ids {
-      display: flex;
-      flex-direction: column;
-      gap: 4px;
-      margin-top: 4px;
-      color: var(--nstrc-profile-text-grey);
-      font-size: 14px;
-      width: 100%;
-    }
-    
-    .nip05-container {
-      display: flex;
-      align-items: center;
-      gap: 4px;
-    }
-    
-    .nip05 {
-      display: flex;
-      align-items: center;
-      gap: 4px;
-    }
-
-    .npub.full {
-      display: inline !important;
-    }
-
-    .npub.masked {
-      display: none !important;
-    }
-
-    .copy-button {
-      display: flex;
-      font-size: 16px;
-      min-width: 15px;
-      min-height: 15px;
-      align-items: center;
-      justify-content: center;
-      border-radius: 5px;
-      cursor: pointer;
-      font-weight: bold;
-      color: var(--nstrc-profile-copy-foreground-color);
-    }
-
-    @media only screen and (max-width: 600px) {
-      button.stat .stat-value {
-        font-size: 18px !important;
-      }
-
-      .npub.full {
-        display: none !important;
-      }
-
-      .npub.masked {
-        display: inline !important;
-      }
-    }
-
-    @media only screen and (max-width: 600px) {
-      :host {
-        --nstrc-follow-btn-padding: 5px 8px !important;
-        --nstrc-follow-btn-font-size: 12px !important;
-        --nstrc-follow-btn-min-height: auto !important;
-        --nstrc-follow-btn-border-radius: 8px !important;
-        --nstrc-follow-btn-error-max-width: 150px !important;
-      }
-    }
-
-  </style>
-  `;
-}
