+++ src/nostr-profile/render.ts
@@ -1,8 +1,8 @@
 import { NDKUserProfile } from '@nostr-dev-kit/ndk';
-import { Theme } from '../common/types';
-import { maskNPub } from '../common/utils';
-import { getProfileStyles } from '../common/theme';
-import { copyIcon } from '../common/icons';
+import { RenderOptions } from '../base/render-options';
+import { renderNpub } from '../base/render-npub';
+import { renderNip05 } from '../base/render-nip05';
+import { renderStats } from './render-stats';
 
 export interface Stats {
   notes: number;
@@ -13,34 +13,37 @@ export interface Stats {
   relays: number;
 }
 
-export interface RenderOptions {
+export interface RenderProfileOptions extends RenderOptions {
   npub: string;
   userProfile: NDKUserProfile;
-  theme: Theme;
-  isLoading: boolean;
   isStatsLoading: boolean;
   isStatsFollowersLoading: boolean;
   isStatsFollowsLoading: boolean;
   stats: Stats;
-  error: string | null;
   showFollow: string | boolean;
   showNpub: boolean;
 }
 
-export function renderProfile(options: RenderOptions): string {
+export function renderProfile(options: RenderProfileOptions): string {
   const {
-    npub,
-    userProfile,
     theme,
     isLoading,
+    isError,
+    errorMessage,
+    npub,
+    userProfile,
     isStatsLoading,
     isStatsFollowersLoading,
     isStatsFollowsLoading,
     stats,
-    error,
     showFollow,
     showNpub,
   } = options;
+
+  if (isError) {
+    return renderError(errorMessage);
+  }
+
 
   // Extract profile data with null checks and default values
   const displayName = userProfile?.displayName || userProfile?.name || '';
@@ -48,36 +51,6 @@ export function renderProfile(options: RenderOptions): string {
   const image = userProfile?.image || '';
   const about = userProfile?.about || '';
   const website = userProfile?.website || '';
-
-
-  if (error) {
-    return `
-      <div class="error-container">
-        <div class="error">!</div>
-        <div class="error-text">${error}</div>
-      </div>
-    `;
-  }
-
-  const styles = getProfileStyles(theme);
-  const maskedNpub = maskNPub(npub);
-
-  const renderNpub = () => {
-    // Convert showNpub to boolean if it's a string
-    const shouldShowNpub = showNpub === true || String(showNpub).toLowerCase() === 'true';
-    if (!shouldShowNpub) return '';
-    
-    return `
-      <div class="npub-container">
-        <div class="npub">
-          <span class="npub-text">${maskedNpub}</span>
-          <span class="copy-button" data-npub="${npub}">
-            ${copyIcon}
-          </span>
-        </div>
-      </div>
-    `;
-  };
 
   const renderFollowButton = () => {
     if (!showFollow) return '';
@@ -90,45 +63,46 @@ export function renderProfile(options: RenderOptions): string {
   };
 
   return `
-    ${styles}
-    <div class="nostr-profile">
-      <div id="profile">
+    <div class="nostr-profile-container">
         <div id="profile_banner">
           ${
             isLoading
               ? '<div style="width: 100%; height: 100%;" class="skeleton"></div>'
               : userProfile.banner
-                ? `<a target="_blank" data-cropped="true" class="profile_image">
+                ? `<a class="profile_image">
                   <img src="${userProfile.banner}" width="524px"/>
                 </a>`
                 : '<div class="banner-placeholder"></div>'
           }
         </div>
+
         <div class="dp_container">
-          <div class="avatar_container">
-            <div class="avatar_wrapper">
-              <div class="xxl_avatar">
-                <div class="backfill">
-                  ${
-                    isLoading
-                      ? '<div style="width: 100%; height: 100%; border-radius: 50%" class="skeleton"></div>'
-                      : `<a target="_blank" data-cropped="true" class="profile_image roundedImage">
-                        <img src="${image}" width="524px" alt="${displayName}" />
-                      </a>`
-                  }
-                </div>
-              </div>
-            </div>
+          <div class="avatar" role="img" aria-label="${displayName}">
+            ${
+              isLoading
+                ? '<div style="width: 100%; height: 100%; border-radius: 50%" class="skeleton"></div>'
+                : `<img
+                    src="${image}"
+                    alt="${displayName}"
+                    width="142" height="142"
+                    loading="lazy" decoding="async"
+                  />`
+            }
           </div>
         </div>
-        
+
+
         <div class="profile_actions">
-          ${renderFollowButton()}
+            ${
+              isLoading ? '<div style="width: 100px; height: 36px; border-radius: 18px;" class="skeleton"></div>'
+              : renderFollowButton()
+            }
         </div>
         
         <div class="profile_data">
           <div class="basic_info">
             <div class="name">
+              <NCTextCopy></NCTextCopy>
               ${
                 isLoading
                   ? '<div style="width: 100px; height: 16px; border-radius: 20px" class="skeleton"></div>'
@@ -137,21 +111,16 @@ export function renderProfile(options: RenderOptions): string {
             </div>
           </div>
           
-          <div class="profile-ids">
-            <div class="nip05-container">
-              ${
-                isLoading
-                  ? '<div style="width: 75px; height: 8px; border-radius: 20px" class="skeleton"></div>'
-                  : nip05
-                    ? `<div class="nip05">
-                        <span>${nip05}</span>
-                        <span id="nip05-copy" class="copy-button">&#x2398;</span>
-                      </div>`
-                    : ''
-              }
-            </div>
-            ${renderNpub()}
-          </div>
+          ${
+            isLoading
+              ? '<div style="width: 75px; height: 12px; border-radius: 20px" class="skeleton"></div>'
+              : renderNip05(nip05)
+          }
+          ${
+            isLoading
+              ? '<div style="width: 75px; height: 12px; border-radius: 20px" class="skeleton"></div>'
+              : renderNpub(npub)
+          }
         </div>
         
         <div class="about">
@@ -174,156 +143,33 @@ export function renderProfile(options: RenderOptions): string {
                 : ''
           }
         </div>
-      </div>
       
       <div class="stats" data-orientation="horizontal">
-        <button class="stat" data-orientation="horizontal">
-          <div class="stat-inner">
-            <div class="stat-value">
-              ${
-                isStatsLoading || isStatsFollowsLoading
-                  ? '<div style="width: 50px; height: 28px; border-radius: 5px" class="skeleton"></div>'
-                  : stats.follows.toLocaleString()
-              }
-            </div>
-            <div class="stat-name">Following</div>
-          </div>
-        </button>
         
-        <button class="stat" data-orientation="horizontal">
-          <div class="stat-inner">
-            <div class="stat-value">
-              ${
-                isStatsLoading || isStatsFollowersLoading
-                  ? '<div style="width: 50px; height: 28px; border-radius: 5px" class="skeleton"></div>'
-                  : stats.followers.toLocaleString()
-              }
-            </div>
-            <div class="stat-name">Followers</div>
-          </div>
-        </button>
+        ${renderStats('Following', stats.follows, isStatsLoading || isStatsFollowsLoading)}
         
-        <button class="stat" data-orientation="horizontal">
-          <div class="stat-inner">
-            <div class="stat-value">
-              ${
-                isStatsLoading
-                  ? '<div style="width: 50px; height: 28px; border-radius: 5px" class="skeleton"></div>'
-                  : stats.notes.toLocaleString()
-              }
-            </div>
-            <div class="stat-name">Notes</div>
-          </div>
-        </button>
+        ${renderStats('Followers', stats.followers, isStatsLoading || isStatsFollowsLoading)}
+
+        ${renderStats('Notes', stats.notes, isStatsLoading)}
         
-        <button class="stat" data-orientation="horizontal">
-          <div class="stat-inner">
-            <div class="stat-value">
-              ${
-                isStatsLoading
-                  ? '<div style="width: 50px; height: 28px; border-radius: 5px" class="skeleton"></div>'
-                  : stats.replies.toLocaleString()
-              }
-            </div>
-            <div class="stat-name">Replies</div>
-          </div>
-        </button>
+        ${renderStats('Replies', stats.replies, isStatsLoading)}
         
-        <button class="stat" data-orientation="horizontal">
-          <div class="stat-inner">
-            <div class="stat-value">
-              ${
-                isStatsLoading
-                  ? '<div style="width: 50px; height: 28px; border-radius: 5px" class="skeleton"></div>'
-                  : stats.zaps.toLocaleString()
-              }
-            </div>
-            <div class="stat-name">Zaps</div>
-          </div>
-        </button>
+        ${renderStats('Zaps', stats.zaps, isStatsLoading)}
+        
       </div>
     </div>
   `;
 }
 
-export function renderLoadingState(theme: Theme): string {
-  const styles = getProfileStyles(theme);
+function renderError(errorMessage: String): string {
   return `
-    ${styles}
-    <div class="nostr-profile">
-      <div id="profile">
-        <div id="profile_banner">
-          <div style="width: 100%; height: 100%;" class="skeleton"></div>
-        </div>
-        <div class="dp_container">
-          <div class="avatar_container">
-            <div class="avatar_wrapper">
-              <div class="xxl_avatar">
-                <div class="backfill">
-                  <div style="width: 100%; height: 100%; border-radius: 50%" class="skeleton"></div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div class="profile_actions">
-          <div style="width: 100px; height: 36px; border-radius: 18px;" class="skeleton"></div>
-        </div>
-        <div class="profile_data">
-          <div class="basic_info">
-            <div class="name">
-              <div style="width: 100px; height: 16px; border-radius: 20px" class="skeleton"></div>
-            </div>
-          </div>
-          <div class="nip05-wrapper" style="display: flex; align-items: flex-start; margin-top: 4px; color: var(--nstr-text-secondary); font-size: 14px; flex-direction: column;">
-            <div class="nip05-container">
-              <div style="width: 75px; height: 8px; border-radius: 20px" class="skeleton"></div>
-              <div class="npub-container" style="width: 100%; margin-top: 8px;">
-                <div class="npub" style="display: flex; align-items: center; gap: 6px; color: var(--nstr-text-secondary); font-family: monospace; font-size: 13px; word-break: break-all;">
-                  <div style="width: 120px; height: 12px; border-radius: 20px; margin-top: 4px;" class="skeleton"></div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div class="about">
-          <div style="width: 100%; height: 12px; border-radius: 20px; margin-bottom: 12px" class="skeleton"></div>
-          <div style="width: 40%; height: 12px; border-radius: 20px" class="skeleton"></div>
-        </div>
-        <div class="links">
-          <div style="width: 150px; height: 12px; border-radius: 20px" class="skeleton"></div>
-        </div>
+    <div class='nostr-profile-container is-error'>
+      <div class='nostr-profile-top-container'>
+        <div class="error-icon">&#9888;</div>
       </div>
-      <div class="stats" data-orientation="horizontal">
-        ${Array(4)
-          .fill(0)
-          .map(
-            () => `
-          <button class="stat" data-orientation="horizontal">
-            <div class="stat-inner">
-              <div class="stat-value">
-                <div style="width: 50px; height: 28px; border-radius: 5px" class="skeleton"></div>
-              </div>
-              <div class="stat-name">
-                <div style="width: 50px; height: 12px; border-radius: 10px; margin: 4px auto 0;" class="skeleton"></div>
-              </div>
-            </div>
-          </button>
-        `
-          )
-          .join('')}
+      <div class='nostr-profile-bottom-container'>
+        ${errorMessage}
       </div>
-    </div>
-  `;
-}
-
-export function renderErrorState(error: string, theme: Theme): string {
-  const styles = getProfileStyles(theme);
-  return `
-    ${styles}
-    <div class="error-container">
-      <div class="error">!</div>
-      <span class="error-text">${error}</span>
     </div>
   `;
 }
